refactor: rename service to new names
